+++ packages/substrate-client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- `getChainSpecData: () => Promise<{name: string, genesisHash: string, properties: any}>`.
+
 ## 0.0.2 - 2024-04-20
 
 ### Fixed